Removed underscores from all packagenames
