+++ src/main/kotlin/io/github/oxisto/reticulated/ast/expression/booleanexpr/BooleanExprVisitor.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package io.github.oxisto.reticulated.ast.expression.boolean_expr
+package io.github.oxisto.reticulated.ast.expression.booleanexpr
 
 import io.github.oxisto.reticulated.ast.CouldNotParseException
 import io.github.oxisto.reticulated.ast.Scope
