+++ src/main/kotlin/io/github/oxisto/reticulated/ast/expression/comparison/Comparison.kt
@@ -18,7 +18,7 @@
 package io.github.oxisto.reticulated.ast.expression.comparison
 
 import io.github.oxisto.reticulated.ast.Node
-import io.github.oxisto.reticulated.ast.expression.boolean_expr.OrExpr
+import io.github.oxisto.reticulated.ast.expression.booleanexpr.OrExpr
 import io.github.oxisto.reticulated.Pair
 
 /**
