+++ src/main/kotlin/io/github/oxisto/reticulated/ast/expression/comparison/ComparisonVisitor.kt
@@ -18,8 +18,8 @@
 package io.github.oxisto.reticulated.ast.expression.comparison
 
 import io.github.oxisto.reticulated.ast.CouldNotParseException
-import io.github.oxisto.reticulated.ast.expression.boolean_expr.BooleanExprVisitor
-import io.github.oxisto.reticulated.ast.expression.boolean_expr.OrExpr
+import io.github.oxisto.reticulated.ast.expression.booleanexpr.BooleanExprVisitor
+import io.github.oxisto.reticulated.ast.expression.booleanexpr.OrExpr
 import io.github.oxisto.reticulated.grammar.Python3BaseVisitor
 import io.github.oxisto.reticulated.grammar.Python3Parser
 import io.github.oxisto.reticulated.Pair
