+++ src/main/kotlin/io/github/oxisto/reticulated/ast/expression/operator/ShiftExpr.kt
@@ -18,7 +18,7 @@
 package io.github.oxisto.reticulated.ast.expression.operator
 
 import io.github.oxisto.reticulated.ast.CouldNotParseException
-import io.github.oxisto.reticulated.ast.expression.boolean_expr.BaseBooleanExpr
+import io.github.oxisto.reticulated.ast.expression.booleanexpr.BaseBooleanExpr
 
 /**
  * This class represents a shift_expr.
