+++ src/test/kotlin/io/github/oxisto/reticulated/SimpleStatementTest.kt
@@ -19,7 +19,7 @@ package io.github.oxisto.reticulated
 
 import io.github.oxisto.reticulated.ast.expression.call.Call
 import io.github.oxisto.reticulated.ast.expression.Identifier
-import io.github.oxisto.reticulated.ast.expression.boolean_ops.OrTest
+import io.github.oxisto.reticulated.ast.expression.booleanops.OrTest
 import io.github.oxisto.reticulated.ast.expression.literal.Integer
 import io.github.oxisto.reticulated.ast.expression.operator.PowerExpr
 import io.github.oxisto.reticulated.ast.simple.AssignmentExpression
