+++ src/test/kotlin/io/github/oxisto/reticulated/expression/ArithmeticTest.kt
@@ -18,7 +18,7 @@
 package io.github.oxisto.reticulated.expression
 
 import io.github.oxisto.reticulated.PythonParser
-import io.github.oxisto.reticulated.ast.expression.boolean_ops.OrTest
+import io.github.oxisto.reticulated.ast.expression.booleanops.OrTest
 import io.github.oxisto.reticulated.ast.expression.comparison.Comparison
 import io.github.oxisto.reticulated.ast.expression.literal.*
 import io.github.oxisto.reticulated.ast.expression.operator.*
