+++ src/test/kotlin/io/github/oxisto/reticulated/expression/ExpressionsTest.kt
@@ -19,7 +19,7 @@ package io.github.oxisto.reticulated.expression
 
 import io.github.oxisto.reticulated.PythonParser
 import io.github.oxisto.reticulated.ast.expression.argument.ArgumentList
-import io.github.oxisto.reticulated.ast.expression.boolean_ops.OrTest
+import io.github.oxisto.reticulated.ast.expression.booleanops.OrTest
 import io.github.oxisto.reticulated.ast.expression.call.Call
 import io.github.oxisto.reticulated.ast.expression.operator.PowerExpr
 import io.github.oxisto.reticulated.ast.simple.ExpressionStatement
